Fix thread makefiles.

+++ Recipe11-2/Listing11-7/main.cpp
@@ -63,9 +63,9 @@ public:
 
         void operator delete(void* pMem)
     {
-        const int index{
+        const intptr_t index{
             (static_cast<MyManagedObject*>(pMem)-s_ManagedObjects.data()) /
-            static_cast<int>(sizeof(MyManagedObject)) };
+            static_cast<intptr_t>(sizeof(MyManagedObject)) };
         if (0 <= index && index < static_cast<int>(s_ManagedObjects.size()))
         {
             s_FreeList.emplace(static_cast<unsigned int>(index));
